+++ iocore/net/quic/QUICStreamManager.cc
@@ -52,7 +52,7 @@ QUICStreamManager::handle_frame(std::shared_ptr<const QUICFrame> frame)
     this->_handle_stream_frame(std::dynamic_pointer_cast<const QUICStreamFrame>(frame));
     break;
   default:
-    Debug(tag, "Unexpected frame type: %02x", frame->type());
+    Debug(tag, "Unexpected frame type: %02x", static_cast<unsigned int>(frame->type()));
     ink_assert(false);
     break;
   }
